+++ pkg/resource/events-logger-secret/grafana-agent-secret.go
@@ -31,12 +31,17 @@ func generateGrafanaAgentSecret(lc loggedcluster.Interface, credentialsSecret *v
 		return nil, errors.WithStack(err)
 	}
 
+	tenant := clusterName
+	if lc.IsCAPI() {
+		tenant = common.DefaultWriteTenant
+	}
+
 	values := values{
 		ExtraSecret: extraSecret{
 			Name: fmt.Sprintf("%s-%s", clusterName, common.GrafanaAgentExtraSecretName()),
 			Data: map[string]string{
 				"logging-url":       fmt.Sprintf(common.LokiURLFormat, lokiURL),
-				"logging-tenant-id": clusterName,
+				"logging-tenant-id": tenant,
 				"logging-username":  writeUser,
 				"logging-password":  writePassword,
 			},