+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use `giantswarm` tenant by default instead of cluster name.
+
 ## [0.16.0] - 2024-11-26
 
 ### Added